requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
discord.py==1.7.3
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-discogs-client==2.3.13
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
